@@ -97,7 +97,7 @@
     "protractor": "~5.4.0",
     "ts-node": "~8.0.0",
     "tslint": "~5.12.0",
-    "typescript": "~3.1.6"
+    "typescript": "~3.2.2"
   },
   "description": "An Ionic project",
   "cordova": {
