@@ -33,14 +33,13 @@ export class ConfigService {
       DB_PASSWORD_VAD10: Joi.string().required(),
       DB_HOST_VAD10: Joi.string().required(),
       DB_PORT_VAD10: Joi.number().required(),
-      UPLOAD_PATH_VAD10: Joi.string().required(),
+      UPLOAD_PATH: Joi.string().required(),
 
       DB_NAME_VAD20: Joi.string().required(),
       DB_USERNAME_VAD20: Joi.string().required(),
       DB_PASSWORD_VAD20: Joi.string().required(),
       DB_HOST_VAD20: Joi.string().required(),
       DB_PORT_VAD20: Joi.number().required(),
-      UPLOAD_PATH_VAD20: Joi.string().required()
 
     });
 
