@@ -1,8 +1,16 @@
-import { Component, OnInit, ViewChild } from '@angular/core';
-import { NavigationService } from '../../../services/navigation.service';
-import { SearchService } from 'src/app/shared/services/search.service';
-import { SharedAnimations } from 'src/app/shared/animations/shared-animations';
-import { Router, RouteConfigLoadStart, ResolveStart, RouteConfigLoadEnd, ResolveEnd } from '@angular/router';
+import { Component, OnInit } from '@angular/core';
+import {
+  NavigationService,
+  SearchService
+} from '@retail/shared';
+
+import {
+  Router,
+  RouteConfigLoadStart,
+  ResolveStart,
+  RouteConfigLoadEnd,
+  ResolveEnd
+} from '@angular/router';
 
 @Component({
   selector: 'app-admin-layout-sidebar-compact',
@@ -10,23 +18,23 @@ import { Router, RouteConfigLoadStart, ResolveStart, RouteConfigLoadEnd, Resolve
   styleUrls: ['./admin-layout-sidebar-compact.component.scss']
 })
 export class AdminLayoutSidebarCompactComponent implements OnInit {
-    moduleLoading: boolean;
-  
-    constructor(
-      public navService: NavigationService,
-      public searchService: SearchService,
-      private router: Router
-    ) { }
-  
-    ngOnInit() {
-      this.router.events.subscribe(event => {
-        if (event instanceof RouteConfigLoadStart || event instanceof ResolveStart) {
-          this.moduleLoading = true;
-        }
-        if (event instanceof RouteConfigLoadEnd || event instanceof ResolveEnd) {
-          this.moduleLoading = false;
-        }
-      });
-    }
+  moduleLoading: boolean;
+
+  constructor(
+    public navService: NavigationService,
+    public searchService: SearchService,
+    private router: Router
+  ) { }
+
+  ngOnInit() {
+    this.router.events.subscribe(event => {
+      if (event instanceof RouteConfigLoadStart || event instanceof ResolveStart) {
+        this.moduleLoading = true;
+      }
+      if (event instanceof RouteConfigLoadEnd || event instanceof ResolveEnd) {
+        this.moduleLoading = false;
+      }
+    });
+  }
 
 }
