@@ -1,8 +1,15 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
-import { NavigationService } from '../../../services/navigation.service';
-import { SearchService } from 'src/app/shared/services/search.service';
-import { SharedAnimations } from 'src/app/shared/animations/shared-animations';
-import { Router, RouteConfigLoadStart, ResolveStart, RouteConfigLoadEnd, ResolveEnd } from '@angular/router';
+import {
+  SearchService,
+  NavigationService
+} from '@retail/shared';
+import {
+  Router,
+  RouteConfigLoadStart,
+  ResolveStart,
+  RouteConfigLoadEnd,
+  ResolveEnd
+} from '@angular/router';
 import { PerfectScrollbarDirective } from 'ngx-perfect-scrollbar';
 
 @Component({
@@ -12,24 +19,24 @@ import { PerfectScrollbarDirective } from 'ngx-perfect-scrollbar';
 })
 export class AdminLayoutSidebarLargeComponent implements OnInit {
 
-    moduleLoading: boolean;
-    @ViewChild(PerfectScrollbarDirective) perfectScrollbar: PerfectScrollbarDirective;
-  
-    constructor(
-      public navService: NavigationService,
-      public searchService: SearchService,
-      private router: Router
-    ) { }
-  
-    ngOnInit() {
-      this.router.events.subscribe(event => {
-        if (event instanceof RouteConfigLoadStart || event instanceof ResolveStart) {
-          this.moduleLoading = true;
-        }
-        if (event instanceof RouteConfigLoadEnd || event instanceof ResolveEnd) {
-          this.moduleLoading = false;
-        }
-      });
-    }
+  moduleLoading: boolean;
+  @ViewChild(PerfectScrollbarDirective) perfectScrollbar: PerfectScrollbarDirective;
+
+  constructor(
+    public navService: NavigationService,
+    public searchService: SearchService,
+    private router: Router
+  ) { }
+
+  ngOnInit() {
+    this.router.events.subscribe(event => {
+      if (event instanceof RouteConfigLoadStart || event instanceof ResolveStart) {
+        this.moduleLoading = true;
+      }
+      if (event instanceof RouteConfigLoadEnd || event instanceof ResolveEnd) {
+        this.moduleLoading = false;
+      }
+    });
+  }
 
 }
