@@ -1,17 +1,18 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { BtnLoadingComponent } from './btn-loading/btn-loading.component';
-import { FeatherIconComponent } from './feather-icon/feather-icon.component';
+import {
+  BtnLoadingComponent,
+  FeatherIconComponent,
+  SearchModule,
+  AsideDockComponent,
+  AsideRightComponent,
+  SharedPipesModule,
+  SharedDirectivesModule,
+  BreadcrumbComponent
+} from '@retail/shared';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
-import { RouterModule } from '@angular/router';
-import { SharedPipesModule } from '../pipes/shared-pipes.module';
-import { SearchModule } from './search/search.module';
-import { SharedDirectivesModule } from '../directives/shared-directives.module';
 import { PerfectScrollbarModule } from 'ngx-perfect-scrollbar';
 import { LayoutsModule } from './layouts/layouts.module';
-import { BreadcrumbComponent } from './breadcrumb/breadcrumb.component';
-import { AsideDockComponent } from './aside-dock/aside-dock.component';
-import { AsideRightComponent } from './aside-dock/aside-right/aside-right.component';
 
 const components = [
   BtnLoadingComponent,
@@ -24,7 +25,6 @@ const components = [
 @NgModule({
   imports: [
     CommonModule,
-    RouterModule,
     LayoutsModule,
     SharedPipesModule,
     SharedDirectivesModule,
