@@ -1,13 +1,25 @@
-import { Injectable } from "@angular/core";
-import { HttpInterceptor, HttpErrorResponse, HttpHandler, HttpRequest, HttpEvent, HttpResponse } from "@angular/common/http";
-import { Observable, throwError } from "rxjs";
+import { Injectable } from '@angular/core';
+import {
+    HttpInterceptor,
+    HttpErrorResponse,
+    HttpHandler,
+    HttpRequest,
+    HttpEvent,
+    HttpResponse
+} from '@angular/common/http';
+import {
+    Observable,
+    throwError
+} from 'rxjs';
 import { tap, catchError, map } from 'rxjs/operators';
 import { ToastrService } from 'ngx-toastr';
-import { Error } from '../../shared/models/error.model';
-import { environment } from '../../../environments/environment';
+import {
+    ServiceOpts,
+    Error
+} from '@retail/shared';
 @Injectable()
 export class ApiFailsInterceptor implements HttpInterceptor {
-    constructor(public toastrService: ToastrService) { }
+    constructor(public toastrService: ToastrService, private options: ServiceOpts) { }
     intercept(request: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
 
         return next.handle(request).pipe(
@@ -18,8 +30,8 @@ export class ApiFailsInterceptor implements HttpInterceptor {
                 return event;
             }),
             catchError((error: HttpErrorResponse) => {
-                const errUrl: string = error.url.substring(0, environment.apiUrl.length);
-                if (errUrl === environment.apiUrl) {
+                const errUrl: string = error.url.substring(0, this.options.apiUrl.length);
+                if (errUrl === this.options.apiUrl) {
                     let data: Error = { reason: '', status: 0 };
 
                     data = {
